Create ee-beta-features directory
The crosswiki_beta report was running in the limn-language-data repository
because of a limitation in reportupdater.  Reportupdater now supports
multiple query folders per repository, so we can clean up.

NOTE: all changes that share this bug should be merged / deployed
together to prevent data loss.  Here is the list of changes for
convenience:

https://gerrit.wikimedia.org/r/#/c/289004/
https://gerrit.wikimedia.org/r/#/c/289005/
https://gerrit.wikimedia.org/r/#/c/289006/
https://gerrit.wikimedia.org/r/#/c/289007/

Bug: T126549
Change-Id: Iec3802574a8fc8dbffefd02b871b089b3f63130e